pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentspec"
version = "0.1.0"
description = "Universal agent manifest standard with resolver — the missing layer between agent definitions and runtimes"
readme = "README.md"
license = "EUPL-1.2"
requires-python = ">=3.11"
authors = [{ name = "Alfonso Sastre" }]
keywords = ["agent", "manifest", "resolver", "llm", "mcp", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "acli-spec>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
agentspec = "agentspec.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentspec"]

[tool.hatch.build.targets.sdist]
include = ["src/agentspec", "schemas", "examples"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B"]
ignore = ["S603", "S607"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
